2-11 Complete
Learned more about forms, and their accompanying input types such as text, email, password, select, checkbox, radio, and textarea types.

In VsCode to input multiple autofill elements, use the following feature `element`+`element`, and it will create both elements in a row for you. For example, this could be used to create an input with a label. Just use type 'label+input' to get the desired result.

In VsCode, if you want multiples of the same elements in a row, use the format `element*n` where n is the number of multiples you want. This can be stacked with ()s to control patterns of which you wish to multiply.