3-5 Learned about CSS units
CSS uses 4 units: px, %, em, and rem

em: is the scale of an object in relation to their parents scale. for example, a font size of 1.25em is the same size as the parents font size multiplied by 1.25.
rem: is the same as em, but it instead it is a relative scale with the html element.