3-6 Text Styling Complete
Learned about a whole bunch of new to me text styling techniques.

font-size: Self explanatory, can be defined using px, em, or rem.
font-weight: Changes the weight of a font. It ranges from 100 to 900 in increments of 100, from thinnest to thickest.
font-style: Changes how a piece of text is emphasized, IE if it should be bolded, or italicized.
font-family: Changes the font family of text, which a secondary option that can be provided if the first family could not be found.
line-height: Changes height of a line of text. Uses a relational scale to the text height and can not be statically defined. A value of 1.8 would add 0.4 font size above the text, and 0.4 font size below the text.
text-transform: Used to define the casing of text, whether it should be all uppercase, all lowercase, capitalised, etc.
letter-spacing: Changes how much space is used between characters. It is good for smaller text.
text-align: Allows text to be aligned to the start(left), center or end(right) of a page.